Move DataOffloadDialog into same dir as DataOffloadButton
